+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - [fuji] fix: should not display the `static` fields in `/fuji inspect configurations` command. (To hides the `not interested` fields.)
 - [core] fix: should remember the `parent GUI` when using the `search function` in `paged GUI`. (This makes the GUI interaction more convenient.)
+- [fuji] feature: now will use different items to represent `argument types` from `vanilla Minecraft` and `fuji`. (For `/fuji inspect argument-types` command)